Add UserCourse and UserAreaOfSpecialization tables
- Updated project file to include new SQL files for UserCourse and UserAreaOfSpecialization.
- Modified QPTemplateStatusTypeId and QPTemplateTagId to use IDENTITY(1,1).
- Removed AreaOfSpecialization and CourseId from User, added collections for UserCourses and UserAreaOfSpecializations.
- Updated ExaminationDbContext to include DbSet for new entities and defined relationships in OnModelCreating.
- Enhanced UserService methods to include related data and changed ID parameter type from int to long.
- Updated IUserService interface to reflect method signature changes.
- Modified Program.cs to add new seeding methods for roles, exam months, QP tags, and designations.
- Created new SQL files and classes for UserAreaOfSpecialization and UserCourse.
- Updated project.nuget.cache to reflect successful NuGet operations.

+++ SKCE.Valuation.API/Examination.Services/ServiceContracts/IUserService.cs
@@ -5,9 +5,9 @@ namespace SKCE.Examination.Services.ServiceContracts
     public interface IUserService
     {
         Task<IEnumerable<User>> GetUsersAsync();
-        Task<User> GetUserByIdAsync(int id);
+        Task<User> GetUserByIdAsync(long id);
         Task<User> AddUserAsync(User user);
         Task<User> UpdateUserAsync(User user);
-        Task<bool> DeleteUserAsync(int id);
+        Task<bool> DeleteUserAsync(long id);
     }
 }
